+++ CHANGELOG.md
@@ -2,5 +2,6 @@
 
 ## Unreleased
 
+- `rtx` core plugins always return ok for `plugin-installed`.
 - Check for existence of `asdf`, then `rtx` for calls to the runtime manager.
 - Initial releaase, beginning with code from `medic-ext-asdf`.